Improve bump chart ranking behavior
Extract BumpChart ranking, annotation, and color helpers into `bumpData.ts`, and tighten label selection so `maxLabels` never exceeds the chart's density budget. The AI/chart-capability updates also ignore non-finite ranking values when evaluating bump-chart fit and use clearer bump-chart description text. Generated API surface and bundle-size snapshots were refreshed for the new BumpChart props and refactor.
