+++ etc/api-surface/semiotic-xy.api.md
@@ -163,6 +163,7 @@ interface-member BumpChartProps::property::enableHover = optional enableHover: b
 interface-member BumpChartProps::property::frameProps = optional frameProps: Partial<Omit<StreamXYFrameProps<Datum>, "chartType" | "customLayout" | "data" | "layoutConfig" | "size">> | undefined
 interface-member BumpChartProps::property::highlightTop = optional highlightTop: number | undefined
 interface-member BumpChartProps::property::hoverHighlight = optional hoverHighlight: "series" | boolean | undefined
+interface-member BumpChartProps::property::labelPriorityAccessor = optional labelPriorityAccessor: ChartAccessor<TDatum, number> | undefined
 interface-member BumpChartProps::property::labelStyle = optional labelStyle: ((datum: TDatum) => React.CSSProperties) | React.CSSProperties | undefined
 interface-member BumpChartProps::property::legend = optional legend: LegendValue
 interface-member BumpChartProps::property::legendInteraction = optional legendInteraction: LegendInteractionMode | undefined
@@ -170,6 +171,7 @@ interface-member BumpChartProps::property::legendPosition = optional legendPosit
 interface-member BumpChartProps::property::lineBy = optional lineBy: ChartAccessor<TDatum, string> | undefined
 interface-member BumpChartProps::property::lineOpacity = optional lineOpacity: number | undefined
 interface-member BumpChartProps::property::lineWidth = optional lineWidth: number | undefined
+interface-member BumpChartProps::property::maxLabels = optional maxLabels: number | undefined
 interface-member BumpChartProps::property::neutralColor = optional neutralColor: string | undefined
 interface-member BumpChartProps::property::pointRadius = optional pointRadius: number | undefined
 interface-member BumpChartProps::property::rankDirection = optional rankDirection: "ascending" | "descending" | undefined
@@ -179,7 +181,7 @@ interface-member BumpChartProps::property::ribbonSizeRange = optional ribbonSize
 interface-member BumpChartProps::property::samplesPerSegment = optional samplesPerSegment: number | undefined
 interface-member BumpChartProps::property::showAxes = optional showAxes: boolean | undefined
 interface-member BumpChartProps::property::showGrid = optional showGrid: boolean | undefined
-interface-member BumpChartProps::property::showLabels = optional showLabels: "both" | "end" | "start" | boolean | undefined
+interface-member BumpChartProps::property::showLabels = optional showLabels: "auto" | "both" | "end" | "start" | boolean | undefined
 interface-member BumpChartProps::property::showLegend = optional showLegend: boolean | undefined
 interface-member BumpChartProps::property::showPoints = optional showPoints: boolean | undefined
 interface-member BumpChartProps::property::styleRules = optional styleRules: StyleRule[] | undefined
