+++ src/components/ai/chartCapabilityTypes.ts
@@ -246,7 +246,7 @@ export interface ChartCapability {
    * If empty, the engine still emits a base suggestion.
    */
   variants?: ReadonlyArray<ChartVariant>
-  /** Caveats independent of variants (e.g. "log scale skipped for negative values"). */
+  /** Caveats that may inspect the active variant (e.g. "log scale skipped for negative values"). */
   caveats?: (profile: ChartDataProfile, variant?: ChartVariant) => ReadonlyArray<string>
   /**
    * Build the props you'd pass to this chart for this dataset. Should produce
